基于redis，rabbitmq和mysql实现异步持久化数据并提升响应速度

+++ src/main/java/com/demo/agent/service/ai/AssistantStream.java
@@ -6,5 +6,5 @@ import dev.langchain4j.service.UserMessage;
 
 public interface AssistantStream{
 
-    TokenStream chat(@MemoryId int memoryId,@UserMessage String message);
+    TokenStream chat(@MemoryId Long memoryId,@UserMessage String message);
 }
